requirements.txt
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0